Implementing the username validation to server

+++ routes/ideas.js
@@ -49,19 +49,26 @@ router.post('/', async (req, res) => {
 
 // Update idea
 router.put('/:id', async (req, res) => {
+   const idea = await Idea.findById(req.params.id);
+   
    try {
-      const updatedIdea = await Idea.findByIdAndUpdate(
-         req.params.id,
-         {
-            $set: {
-               text: req.body.text,
-               tag: req.body.tag
-            }
-         },
-         {new: true}
-      )
-      res.json({success: true, message: 'Idea Successfully Updated', data: updatedIdea})
-   }
+      //Match the usernames
+      if(idea.username === req.body.username) {
+         const updatedIdea = await Idea.findByIdAndUpdate(
+            req.params.id,
+            {
+               $set: {
+                  text: req.body.text,
+                  tag: req.body.tag
+               }
+            },
+            {new: true}
+         );
+         return res.json({success: true, message: 'Idea Successfully Updated', data: updatedIdea});
+      }
+       // Usernames do not match
+       res.status(403).json({success: false, error: 'You are not authorized to update this account'});    
+     }
    catch (error) {
       console.log(error);
       res.status(500).json({success: false, error: 'Something went wrong'});
@@ -70,8 +77,15 @@ router.put('/:id', async (req, res) => {
 
 router.delete('/:id', async (req, res) => {
    try {
-      await Idea.findByIdAndDelete(req.params.id);
-      res.json({success: true, message: 'Idea Successfully Deleted', data: {}})
+      const idea = await Idea.findById(req.params.id);
+      // Match the usernames
+      if(idea.username === req.body.username) {
+         await Idea.findByIdAndDelete(req.params.id);
+         res.json({success: true, message: 'Idea Successfully Deleted', data: {}})
+      }
+
+      // Usernames do not match
+      res.status(403).json({success: false, error: 'You are not authorized to delete this account'});
    } 
    catch (error) {
       console.log(error);
